requirements.txt
Click==7.0
demo==0.1
Flask==1.1.1
Flask-SQLAlchemy==2.4.1
itsdangerous==1.1.0
Jinja2==2.11.0
MarkupSafe==1.1.1
python-dotenv==0.10.5
spotify-demo==0.1
SQLAlchemy==1.3.13
Werkzeug==0.16.1
